requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
attrs==24.2.0
biopython==1.79
Bottleneck==1.3.7
brotlipy==0.7.0
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.0
datasets==2.19.1
dill==0.3.8
filelock==3.13.1
frozenlist==1.4.0
fsspec==2024.3.1
# gmpy2==2.1.0a2
einops==0.8.0
huggingface_hub==0.24.6
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.3
mkl-fft==1.3.11
mkl-random==1.2.8
mkl-service==2.4.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.3
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pip==24.2
pyarrow==16.1.0
pycparser==2.21
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.12.0
setuptools==75.1.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.11.0
urllib3==1.26.8
wheel==0.44.0
xxhash==2.0.2
yarl==1.11.0
prokbert==0.0.44
openalexapi==0.0.1a0 
seaborn==0.13.2

